+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- return status code 1 when bit test has failing tests
+
 ## [13.0.5-dev.17 - 2018-10-21]
 
 - update bit-javascript to support Vue files with typescript